# add delete state of row block v2 for performance