--- a/be/src/olap/generic_iterators.cpp
+++ b/be/src/olap/generic_iterators.cpp
@@ -94,6 +94,7 @@ Status AutoIncrementIterator::next_batch(RowBlockV2* block) {
     }
     block->set_num_rows(row_idx);
     block->set_selected_size(row_idx);
+    block->set_delete_state(DEL_PARTIAL_SATISFIED);
     if (row_idx > 0) {
         return Status::OK();
     }
@@ -138,6 +139,8 @@ public:
     // Only when this function return true, current_row()
     // will return a valid row
     bool valid() const { return _valid; }
+
+    int is_partial_delete() { return _block.delete_state() == DEL_PARTIAL_SATISFIED; }
 
 private:
     // Load next block into _block
@@ -254,6 +257,7 @@ Status MergeIterator::init(const StorageReadOptions& opts) {
 
 Status MergeIterator::next_batch(RowBlockV2* block) {
     size_t row_idx = 0;
+    block->set_delete_state(DEL_NOT_SATISFIED);
     for (; row_idx < block->capacity() && !_merge_heap->empty(); ++row_idx) {
         auto ctx = _merge_heap->top();
         _merge_heap->pop();
@@ -262,6 +266,9 @@ Status MergeIterator::next_batch(RowBlockV2* block) {
         // copy current row to block
         copy_row(&dst_row, ctx->current_row(), block->pool());
 
+        if (ctx->is_partial_delete()) {
+            block->set_delete_state(DEL_PARTIAL_SATISFIED);
+        }
         RETURN_IF_ERROR(ctx->advance());
         if (ctx->valid()) {
             _merge_heap->push(ctx);