--- a/be/src/olap/row_block2.h
+++ b/be/src/olap/row_block2.h
@@ -62,6 +62,7 @@ public:
         for (int i = 0; i < _selected_size; ++i) {
             _selection_vector[i] = i;
         }
+        _delete_state = DEL_SATISFIED;
     }
 
     // convert RowBlockV2 to RowBlock
@@ -92,6 +93,14 @@ public:
         _selected_size = selected_size;
     }
 
+    uint8_t delete_state() const { return _delete_state; }
+
+    void set_delete_state(uint8_t delete_state) {
+        if (delete_state > _delete_state) {
+            _delete_state = delete_state;
+        }
+    }
+
 private:
     Schema _schema;
     size_t _capacity;
@@ -112,6 +121,9 @@ private:
     uint16_t* _selection_vector;
     // selected rows number
     uint16_t _selected_size;
+
+    // block delete state
+    uint8_t _delete_state;
 };
 
 // Stands for a row in RowBlockV2. It is consisted of a RowBlockV2 reference