--- a/be/src/olap/rowset/segment_v2/column_reader.h
+++ b/be/src/olap/rowset/segment_v2/column_reader.h
@@ -149,10 +149,14 @@ public:
     ColumnIterator() { }
     virtual ~ColumnIterator() { }
 
+<<<<<<< HEAD
     virtual Status init(const ColumnIteratorOptions& opts) {
         _opts = opts;
         return Status::OK();
     }
+=======
+    virtual Status init() { return Status::OK(); }
+>>>>>>> Add default value column iterator #1834 (#1835)
 
     // Seek to the first entry in the column.
     virtual Status seek_to_first() = 0;