… be/src/olap/rowset/segment_v2/segment.h
… be/src/olap/rowset/segment_v2/segment_iterator.cpp
… be/src/olap/rowset/segment_v2/segment_iterator.h
… be/test/olap/rowset/segment_v2/segment_test.cpp