1. BetaRowset do not set block_status of RowBlock corrently, which will incur more compute cost in reader. This patch will set block_status for BetaRowset.
2. the rowset_meta() of Rowset return shared_ptr reference instead of shared_ptr object for performance
3. related issue #2040