@kangpinghuang 2019-10-24
this improment will reduce my test query time cost by 200ms(total 1.2s).

my test case:
table: compreh_interest
model: dup
record number:14645005
query:
select count(*) from compreh_interest;

original time: 1.2s
optimized time: 1.0s